ARM: dts: msm8998: remove some of the higher inefficient frequencies
The inconsistency in number of data removed from frequency table and EM is
because as far as I understand, Google's EM only accounts from frequencies
from 300->2361 from the big cluster and not anything beyond that.
So, it's safe to remove only the ones accounted for, ie. for
2342400 kHz and 2361600 kHz.

Signed-off-by: idkwhoiam322 <snkm1999@gmail.com>
Do not remove until 2361.6MHz and their EM data
Signed-off-by: ederekun <sedrickvince@gmail.com>
Signed-off-by: dotkit <ewprjkt@proton.me>